backend/business-server/requirements.txt
Django==3.2.3
PyMySQL==1.0.2
djangorestframework==3.12.4
